Several changes
Use the global variable sessionsettings_config_file in all places that
must be used.

QSettings::setPath was already setted so it's not needed. Deleted.

Instead using variable for date and time format in case user settings
aren't setted, now we go to Qt::DefaultLocaleShortDate. To perform this
operation, we must work with Date and Time separately. Now
DirWidget::date_format is an QStringList, first item for date and second
for time. The QDateTime with format "yyyyMMddhhmmsszzz" is stored on
whatsThis variable for being used in sort operations.

CQTreeWidgetItem operator< function has been simplified. Now, to check
dates, we used the value stored in whatsThis.