Another large update for Lumina 2: Get the new LShortcutEvents class setup and functional. This is what converts the key presses and mouse clicks into a format that can be scanned for user-defined shortcuts to do things in the session. NOTES: The format of the keys.conf files uses "keycodes" (in ascending order") delimited by "+" to make the settings searchable and non-ordered-specific. Example: [strict] 22+37+64=logout
This is a shortcut for [L_Ctrl (37), L_Alt (64), Backspace (22)] and it will trigger the start of the logout procedures.
The "strict" section (and it's alternative "desktop" section) determine priority of the shortcut (strict is always evaluated, desktop is only evaluated if the desktop has focus and there is no corresponding "strict" shortcut).